moved googleStrategy file to passport folder
